refactor: app选择按钮重命名为模型选择
