+++ components/chat-input/chat-input.tsx
@@ -15,7 +15,7 @@ import { ChatButton } from "./button"
 import { ChatTextInput } from "./text-input"
 import { ChatContainer } from "./container"
 import { ChatButtonArea, ChatTextArea } from "./layout"
-import { AppSelectorButton } from "./app-selector-button"
+import { AppSelectorButton } from "./model-selector-button"
 import { create } from "zustand"
 import { TooltipWrapper } from "@components/ui/tooltip-wrapper"
 import { uploadDifyFile } from "@lib/services/dify/file-service"
